Prototype of using active_job
* if/else approach for starter
* a lot of code smells now visible which can be a guideline for a
  upcoming refactorings.

It is now more obvious which instance variables are necessary and which
are not for both strategies.

I am not sure which place is best for this setup

      case @config.strategy
      when :thread
        @queue = Queue.new
        @worker_mutex = Mutex.new
        @worker_thread = nil
      when :active_job
        require 'chillout/job'
        Chillout::Job.dispatcher = @dispatcher
      end

I decided not to move it into #start because then client with :thread
strategy could not be used without calling #start.

Railtie was calling #start_worker but this is not necessary for
active_job strategy so I changed it to call #start which does nothing
for :active_job strategy.

Now starts look like an anti-pattern:

    def start
      case @config.strategy
      when :thread
        start_worker
      when :active_job
      end
    end

But maybe our worker thread should be always lazy-loaded and we should
remove #start method and not call it in railtie?

I am now leaning towards a reflaction that we need enqueueing an
dequeueing strategy.

I wasn't sure if I can use a global dispatcher like this:

Chillout::Job.dispatcher = @dispatcher

but all the dependencies are frozen which indicates they are state-less
which indicates they are thread-safe as well. It might matter in case of
active_job adapter such as sidekiq which could use multiple threads and
send multiple metrics at the same time.

@http_client = HttpClient.new(@config, logger).freeze
@event_data_builder = EventDataBuilder.new(@config).freeze
@server_side = ServerSide.new(@event_data_builder, @http_client).freeze
@dispatcher = Dispatcher.new(@server_side).freeze

Thanks for reading ;)