Sidekiq integration detection
Use same code in testing to enable/disable it.

Sidekiq has a similar notion of middleware to Rack: these are small bits of code that can implement functionality. Sidekiq breaks middleware into client-side and server-side.

* Server-side middleware runs 'around' job processing.
* Client-side middleware runs before the pushing of the job to Redis and allows you to modify/stop the job before it gets pushed.

I decided to go with

  defined?(::Sidekiq) && ::Sidekiq.server?

to checking if we should enable our sidekiq middleware in a same way
between testing (gem) and production (rails app)

  Sidekiq.server_middleware.add Middleware::SidekiqCreationsMonitor
  Sidekiq::Testing.server_middleware.add Middleware::SidekiqCreationsMonitor

because Sidekiq::Testing does have the #configure_server method of
that Sidekiq does:

  Sidekiq.configure_server do |config|
    config.server_middleware do |chain|
      chain.add Middleware
    end
  end

so instead of detecting that we are running in a sidekiq server with
the #configure_server(&proc) method which is just a nice sugar

  def self.configure_server
    yield self if server?
  end

I went with the mentioned

  defined?(::Sidekiq) && ::Sidekiq.server?

which does the. Now we can use the same piece of code in
both places (rails app with sidekiq and testing in the gem)
for sidekiq integration initialization

  def enable(client, a_module = ::Sidekiq)
    require 'chillout/middleware/sidekiq'
    @module = a_module
    @module.server_middleware.add Middleware::SidekiqCreationsMonitor,
      client: client
  end

I wanted to pass the default module (a_module = ::Sidekiq) in
initializer but that would mean the constant would be evaluated before
the method #available? is called which would counter the whole reason
for having this call being able to easily check if we need an
integration or not.