Prevent metrics queue from growining infinitelly
I considered many approaches for it and decided to go with this one
because of its simplicity.

I am not sure about the number 5K but this can always be
configured/changed later.

I considerd using SizedQueue ( http://ruby-doc.org/stdlib-2.0.0/libdoc/thread/rdoc/SizedQueue.html )
but it has a blocking interface when the queue is full which is a big
NO NO.

I considered using Concurrent::Array (
http://ruby-concurrency.github.io/concurrent-ruby/Concurrent/Array.html
) but that won't let us easily have blocking queue.pop which can wait
for anything being published on the queue.

I considered approaches in which the worker would be taking max 5K of
metrics and cleared the rest. But during the time when worker is busy
sending metrics the queue could grow significantly anyway. So I decided
against it as well.

I believe we should change get_all_containers_to_process to limit number
of taken metrics as well because if they are being added sufficiently
quickly it would keep dequeueing them infinitelly as well. This will be
next thing to do.