Ability to disable tracking some metrics out of box
reflections/wonders:

a) Some metrics such as creations/custom metrics are accumulated (in
Chillout.creations) over a period of a request/job
and then enqueued in the client.

b) Other metrics such as controller performance metrics are directly
enqueued in the client.

Since the client already knows how to squeeze metrics into one request,
once can wonder if strategy a) is necessary. However the good side is
that we never block on passing the metric to clients shared queue during
the request in such case.

I also wonder which strategy to choose for other metrics such as SQL
queries.

I am not sure how to test railtie in gem tests right now. Awaiting
further insights.