Metrics with tags and multiple measurements
Custom metrics like that cannot be easily merged together like in the
case of creations metrics.

Adding the metrics only enqueues them, without sending
immediately.

I considered using ActiveSupport::Notifications but that would just be
another, unnecessary layer of indirection as subscriber would need to
re-queue to Client#queue. For long time I feared that the client is
going to become a global and it seems that now it happened.

Alternatively we could extract custom SizedQueue class with proper API
that we like and expose just it as global.

Right now it is hard to exactly say what's the responsibility of Client.
Where it starts and where it ends. But the code is still good enough to
understand so let's leave it like that for now.