Integrate with state_machine
Integration is enabled after eager loading code so that we can easily
access all ancestors of ActiveRecord::Base and check their configured
state_machine integrations.

We manually add our callback as first in collection using:

  state_machine.callbacks[:before].unshift

instead of

  state_machine.around_transition do |instance, transition, block|

because we want our tracking callback to be outermost instead of
innermost to track performance of all inner hooks and callbacks.